// src/app/core/models/gym.models.ts
export interface RoutineExercise {
  id?: string;
  order?: number;
  name: string;
  sets: number;
  reps: string;
  rest: string;
}

export interface RoutineDay {
  id?: string;
  dayOrder?: number;
  name: string;
  exercises: RoutineExercise[];
}

export interface UpdateRoutineRequest {
  days: RoutineDay[];
}